Added class AspectResolver
