+++ src/AOP/AspectDispatcher.php
@@ -15,15 +15,18 @@
 	
 	class AspectDispatcher {
 		
+		private AnnotationReader $annotationReader;
+		private Container $di;
+		
 		/**
 		 * AspectDispatcher constructor
 		 * @param AnnotationReader $annotationReader
 		 * @param Container $di
 		 */
-		public function __construct(
-			private readonly AnnotationReader $annotationReader,
-			private readonly Container $di
-		) {}
+		public function __construct(AnnotationReader $annotationReader, Container $di) {
+			$this->annotationReader = $annotationReader;
+			$this->di = $di;
+		}
 		
 		/**
 		 * Dispatch a call
@@ -38,16 +41,16 @@
 		public function dispatch(Request $request, object $controller, string $method, array $arguments): Response {
 			// Create method context
 			$context = new MethodContext(
+				request: $request,
 				target: $controller,
 				methodName: $method,
 				arguments: $arguments,
 				reflection: new \ReflectionMethod($controller, $method),
-				annotations: $this->annotationReader->getMethodAnnotations($controller, $method),
-				request: $request
+				annotations: $this->annotationReader->getMethodAnnotations($controller, $method)
 			);
 			
 			// Get and instantiate aspects
-			$aspects = $this->resolveAspects($controller, $method);
+			$aspects = $this->getResolvedAspects($controller, $method);
 			
 			// Execute request aspects
 			foreach ($aspects as $aspect) {
@@ -83,53 +86,30 @@
 			
 			return $result;
 		}
-		
 		/**
-		 * Resolves all aspects that should be applied to a controller method
-		 * Combines class-level aspects (applied to all methods) with method-level aspects
-		 * @param object $controller The controller instance
-		 * @param string $method The method name being called
-		 * @return array Array of aspect annotation instances
+		 * Resolves and instantiates aspects for a given controller method.
+		 * @param object $controller The controller instance that contains the target method
+		 * @param string $method The name of the method to resolve aspects for
+		 * @return array Array of instantiated aspect objects ready for execution
 		 */
-		private function resolveAspects(object $controller, string $method): array {
-			// Get all annotations from the controller class and filter for aspect annotations
-			// Class-level aspects apply to all methods in the controller
-			$classAnnotations = array_filter($this->annotationReader->getClassAnnotations($controller), function ($e) {
-				return $e instanceof AspectAnnotation;
-			});
+		private function getResolvedAspects(object $controller, string $method): array {
+			// Instantiate AspectResolver, which is used to find AOP classes
+			$aspectResolver = new AspectResolver($this->annotationReader);
 			
-			// Get all annotations from the specific method and filter for aspect annotations
-			// Method-level aspects only apply to this specific method
-			$methodAnnotations = array_filter($this->annotationReader->getMethodAnnotations($controller, $method), function ($e) {
-				return $e instanceof AspectAnnotation;
-			});
+			// Use the aspect resolver to determine which aspects apply to this controller method
+			// Returns an array where keys are aspect class names and values are constructor parameters
+			$aspects = $aspectResolver->resolve($controller, $method);
 			
-			// Merge class-level and method-level aspects
-			// Class aspects are applied first, then method aspects
-			// This allows method-level aspects to override or extend class-level behavior
-			$allAnnotations = array_merge($classAnnotations, $methodAnnotations);
-			
-			// Convert annotation instances to actual aspect objects with their parameters
-			$aspects = [];
-			
-			foreach ($allAnnotations as $annotation) {
-				// Extract the aspect class name from the 'value' parameter
-				// This comes from @InterceptWith(CacheAspect::class, ttl=300)
-				$aspectClass = $annotation->getInterceptClass(); // e.g., CacheAspect::class
-				
-				// Get all annotation parameters except 'value' to pass to aspect constructor
-				// For @InterceptWith(CacheAspect::class, ttl=300), this gives us ['ttl' => 300]
-				$parameters = array_filter($annotation->getParameters(), function ($key) {
-					return $key !== 'value';
-				}, ARRAY_FILTER_USE_KEY);
-				
-				// Instantiate the aspect class through DI container
-				// DI will autowire dependencies while using annotation parameters for constructor
-				// Example: new CacheAspect(cacheService: $injectedCache, ttl: 300)
-				$aspects[] = $this->di->get($aspectClass, $parameters);
+			// Iterate through each resolved aspect and instantiate it
+			$aspectsResolved = [];
+
+			foreach($aspects as $aspectClass => $parameters) {
+				// Use dependency injection container to create aspect instance with its parameters
+				// The DI container handles constructor injection and aspect lifecycle
+				$aspectsResolved[] = $this->di->get($aspectClass, $parameters);
 			}
 			
-			return $aspects;
+			return $aspectsResolved;
 		}
 		
 		/**
