+++ src/Kernel.php
@@ -116,6 +116,7 @@
 
 				// Create aspect-aware dispatcher
 			    $aspectDispatcher = new AspectDispatcher($this->annotationsReader, $this->di);
+				
 			    return $aspectDispatcher->dispatch(
 					$request,
 				    $controller,
